+++ CHANGELOG.md
@@ -8,6 +8,15 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
+## [0.0.7] - Dev branch
+
+### Changed
+
+- Bump `@types/node` to v.25.9.3.
+- Refactor `test-publish` to allow also state update.
+
+<a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
+
 ## [0.0.6] - 2026-06-10
 
 ### Added